+++ Percursus.lua
@@ -90,7 +90,7 @@ end
 function percursusFrame:QUEST_ACCEPTED(_, questID)
     local result = GetRaceData(questID)
 
-    --Utils:PrintDebug("questID: " .. questID)
+    Utils:PrintDebug("questID: " .. questID)
 
     if result ~= nil then
         Utils:PrintDebug("Event 'QUEST_ACCEPTED' fired. Payload: " .. C_QuestLog.GetTitleForQuestID(questID) .. " (" .. questID ..")")
@@ -165,7 +165,7 @@ GossipFrame:HookScript("OnShow",function()
 		local npcID = select(6, strsplit("-", tostring(UnitGUID("target"))))
         npcID = tonumber(npcID)
 
-        --skyridingRaceTracker:PrintDebug("npcID: " .. npcID)
+        --Utils:PrintDebug("npcID: " .. npcID)
 
         if raceDataTable[npcID] ~= nil then
             RaceTimeOverview:ShowRaceOverview(npcID)