+++ data/event/EasternKingdomCupData.lua
@@ -1,0 +1,121 @@
+local _, PER = ...
+
+if PER.RACE_DATA == nil then
+    PER.RACE_DATA = {}
+end
+
+local raceDataTable = PER.RACE_DATA
+
+-- Eastern Kingdom Cup
+-- Gilneas Gambit R1 (XXX) - R-11.2.0
+raceDataTable[206246] = {
+    010201, 13, {
+        NORMAL            = {76309, 2536, 0, 78, 83},
+        ADVANCED          = {76310, 2552, 0, 74, 79},
+        REVERSE           = {76311, 2568, 0, 74, 79}
+    }
+}
+-- Loch Modan Loop R2 (XXX) - R-11.2.0
+raceDataTable[206461] = {
+    010202, 13, {
+        NORMAL            = {76339, 2537, 0, 63, 68},
+        ADVANCED          = {76340, 2553, 0, 61, 66},
+        REVERSE           = {76341, 2569, 0, 63, 68}
+    }
+}
+-- Searing Slalom R3 (XXX) - R-11.2.0
+raceDataTable[206618] = {
+    010203, 13, {
+        NORMAL            = {76357, 2538, 0, 52, 57},
+        ADVANCED          = {76358, 2554, 0, 46, 51},
+        REVERSE           = {76359, 2570, 0, 43, 48}
+    }
+}
+-- Twilight Terror R4 (XXX) - R-11.2.0
+raceDataTable[206764] = {
+    010204, 13, {
+        NORMAL            = {76364, 2539, 0, 73, 78},
+        ADVANCED          = {76365, 2555, 0, 68, 73},
+        REVERSE           = {76366, 2571, 0, 66, 71}
+    }
+}
+-- Deadwind Derby R5 (Totenwindtour) - R-11.2.0
+raceDataTable[206823] = {
+    010205, 13, {
+        NORMAL            = {76380, 2540, 413966, 60, 65}, --XXX
+        ADVANCED          = {76381, 2556, 413967, 59, 64},
+        REVERSE           = {76382, 2572, 413968, 59, 64}
+    }
+}
+-- Elwynn Forest Flash R6 (Elwynnetappe) - R-11.2.0
+raceDataTable[206893] = {
+    010206, 13, {
+        NORMAL            = {76397, 2541, 414016, 73, 78}, --XXX
+        ADVANCED          = {76399, 2557, 414017, 66, 71},
+        REVERSE           = {76400, 2573, 414018, 63, 68}
+    }
+}
+-- Gurubashi Gala R7 (Gurubashigala) - R-11.2.0
+raceDataTable[207163] = {
+    010207, 13, {
+        NORMAL            = {76438, 2542, 414349, 56, 61}, --XXX
+        ADVANCED          = {76439, 2558, 414350, 49, 54},
+        REVERSE           = {76440, 2574, 414351, 50, 55}
+    }
+}
+-- Ironforge Interceptor R8 (XXX) - R-11.2.0
+raceDataTable[207173] = {
+    010208, 13, {
+        NORMAL            = {76445, 2543, 0, 70, 75},
+        ADVANCED          = {76446, 2559, 0, 64, 69},
+        REVERSE           = {76447, 2575, 0, 60, 65}
+    }
+}
+-- Blasted Lands Bolt R9 (Funke der Verwüsteten Lande) - R-11.2.0
+raceDataTable[207357] = {
+    010209, 13, {
+        NORMAL            = {76469, 2544, 414616, 69, 74}, --XXX
+        ADVANCED          = {76470, 2560, 414617, 62, 67},
+        REVERSE           = {76471, 2576, 414618, 64, 69}
+    }
+}
+-- Plaguelands Plunge R10 (XXX) - R-11.2.0
+raceDataTable[207448] = {
+    010210, 13, {
+        NORMAL            = {76510, 2545, 0, 63, 68},
+        ADVANCED          = {76512, 2561, 0, 53, 58},
+        REVERSE           = {76513, 2577, 0, 58, 63}
+    }
+}
+-- Booty Bay Blast R11 (Beutebuchtgeballer) - R-11.2.0
+raceDataTable[207455] = {
+    010211, 13, {
+        NORMAL            = {76515, 2546, 414751, 63, 68}, --XXX
+        ADVANCED          = {76516, 2562, 414755, 57, 62},
+        REVERSE           = {76517, 2578, 414756, 56, 61}
+    }
+}
+-- Fuselight Night Flight R12 (XXX) - R-11.2.0
+raceDataTable[207467] = {
+    010212, 13, {
+        NORMAL            = {76523, 2547, 0, 64, 69},
+        ADVANCED          = {76524, 2563, 0, 58, 63},
+        REVERSE           = {76525, 2579, 0, 58, 63}
+    }
+}
+-- Krazzworks Klash R13 (XXX) - R-11.2.0
+raceDataTable[207485] = {
+    010213, 13, {
+        NORMAL            = {76527, 2548, 0, 71, 76},
+        ADVANCED          = {76528, 2564, 0, 64, 69},
+        REVERSE           = {76529, 2580, 0, 62, 67}
+    }
+}
+-- Redridge Rally R14 (Rotkammrally) - R-11.2.0
+raceDataTable[207524] = {
+    010214, 13, {
+        NORMAL            = {76536, 2549, 414891, 57, 62}, --XXX
+        ADVANCED          = {76537, 2565, 414892, 52, 57},
+        REVERSE           = {76538, 2581, 414893, 52, 57}
+    }
+}